+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageContainerView.swift
@@ -28,6 +28,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
     @State private var frame: CGRect = .zero
     @State private var computeFrame = false
     @State private var offsetX: CGFloat = 0
+    @State private var offsetYAvatar: CGFloat = 0
     @GestureState private var offset: CGSize = .zero
 
     private let replyThreshold: CGFloat = 60
@@ -70,6 +71,8 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                             for: utils.messageCachingUtils.authorInfo(from: message)
                         )
                         .opacity(showsAllInfo ? 1 : 0)
+                        .offset(y: bottomReactionsShown ? offsetYAvatar : 0)
+                        .animation(nil)
                     }
                 }
 
@@ -77,7 +80,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                     if isMessagePinned {
                         MessagePinDetailsView(
                             message: message,
-                            reactionsShown: reactionsShown
+                            reactionsShown: topReactionsShown
                         )
                     }
 
@@ -90,7 +93,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                     )
                     .overlay(
                         ZStack {
-                            reactionsShown ?
+                            topReactionsShown ?
                                 factory.makeMessageReactionView(
                                     message: message,
                                     onTapGesture: {
@@ -172,6 +175,29 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                         .accessibilityElement(children: .contain)
                         .accessibility(identifier: "MessageRepliesView")
                     }
+                    
+                    if bottomReactionsShown {
+                        factory.makeBottomReactionsView(message: message, showsAllInfo: showsAllInfo) {
+                            handleGestureForMessage(
+                                showsMessageActions: false,
+                                showsBottomContainer: false
+                            )
+                        } onLongPress: {
+                            handleGestureForMessage(showsMessageActions: false)
+                        }
+                        .background(
+                            GeometryReader { proxy in
+                                let frame = proxy.frame(in: .local)
+                                let height = frame.height
+                                Color.clear.preference(key: HeightPreferenceKey.self, value: height)
+                            }
+                        )
+                        .onPreferenceChange(HeightPreferenceKey.self) { value in
+                            if value != 0 {
+                                self.offsetYAvatar = -(value ?? 0)
+                            }
+                        }
+                    }
 
                     if showsAllInfo && !message.isDeleted {
                         if message.isSentByCurrentUser && channel.config.readEventsEnabled {
@@ -210,7 +236,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
         }
         .padding(
             .top,
-            reactionsShown && !isMessagePinned ? messageListConfig.messageDisplayOptions.reactionsTopPadding(message) : 0
+            topReactionsShown && !isMessagePinned ? messageListConfig.messageDisplayOptions.reactionsTopPadding(message) : 0
         )
         .padding(.horizontal, messageListConfig.messagePaddings.horizontal)
         .padding(.bottom, showsAllInfo || isMessagePinned ? paddingValue : 2)
@@ -247,6 +273,20 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
         messageListConfig.messageDisplayOptions.spacerWidth(width ?? 0)
     }
 
+    private var topReactionsShown: Bool {
+        if messageListConfig.messageDisplayOptions.reactionsPlacement == .bottom {
+            return false
+        }
+        return reactionsShown
+    }
+    
+    private var bottomReactionsShown: Bool {
+        if messageListConfig.messageDisplayOptions.reactionsPlacement == .top {
+            return false
+        }
+        return reactionsShown
+    }
+    
     private var reactionsShown: Bool {
         !message.reactionScores.isEmpty
             && !message.isDeleted
@@ -289,7 +329,10 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
         }
     }
 
-    private func handleGestureForMessage(showsMessageActions: Bool) {
+    private func handleGestureForMessage(
+        showsMessageActions: Bool,
+        showsBottomContainer: Bool = true
+    ) {
         computeFrame = true
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
             computeFrame = false
@@ -300,7 +343,8 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                     frame: frame,
                     contentWidth: contentWidth,
                     isFirst: showsAllInfo,
-                    showsMessageActions: showsMessageActions
+                    showsMessageActions: showsMessageActions,
+                    showsBottomContainer: showsBottomContainer
                 )
             )
         }
@@ -331,6 +375,7 @@ public struct MessageDisplayInfo {
     public let contentWidth: CGFloat
     public let isFirst: Bool
     public var showsMessageActions: Bool = true
+    public var showsBottomContainer: Bool = true
     public var keyboardWasShown: Bool = false
 
     public init(
@@ -339,6 +384,7 @@ public struct MessageDisplayInfo {
         contentWidth: CGFloat,
         isFirst: Bool,
         showsMessageActions: Bool = true,
+        showsBottomContainer: Bool = true,
         keyboardWasShown: Bool = false
     ) {
         self.message = message
@@ -347,5 +393,6 @@ public struct MessageDisplayInfo {
         self.isFirst = isFirst
         self.showsMessageActions = showsMessageActions
         self.keyboardWasShown = keyboardWasShown
+        self.showsBottomContainer = showsBottomContainer
     }
 }