+++ StreamChatSwiftUITests/Tests/Utils/ViewFactory_Tests.swift
@@ -803,6 +803,22 @@ class ViewFactory_Tests: StreamChatTestCase {
         // Then
         XCTAssert(modifier is EmptyViewModifier)
     }
+    
+    func test_viewFactory_makeBottomReactionsView() {
+        // Given
+        let viewFactory = DefaultViewFactory.shared
+        
+        // When
+        let view = viewFactory.makeBottomReactionsView(
+            message: .mock(),
+            showsAllInfo: true,
+            onTap: {},
+            onLongPress: {}
+        )
+        
+        // Then
+        XCTAssert(view is BottomReactionsView)
+    }
 }
 
 extension ChannelAction: Equatable {